refactor: Move the factory functions to a dedicated module
- Move the in-memory dataloader and from_* factories to _base/_factory.py
- Relocate the _from_components engine helper alongside them
- Keep _dataloader.py focused on the abstract base and its engine
